[CL] Extract create aqueous solution function.
